requirements.txt
asgiref==3.7.2
autopep8==2.0.4
cffi==1.16.0
cryptography==42.0.2
distlib==0.3.8
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
gunicorn==21.2.0
mysqlclient==2.2.4
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pycodestyle==2.11.1
pycparser==2.21
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.25.0
whitenoise==6.6.0
